Complete overhaul of weapon system, attacks are now decoupled from weapon, attack system is fully modular
