+++ Game/ai_boss/scenes/dagger.tscn
@@ -1,4 +1,4 @@
-[gd_scene load_steps=17 format=3 uid="uid://cugyiq1gej4hh"]
+[gd_scene load_steps=17 format=3 uid="uid://ir5gnufcircy"]
 
 [ext_resource type="Script" uid="uid://cjemx5v6p4ud" path="res://scripts/MeleeWeapon.cs" id="1_5mald"]
 [ext_resource type="Texture2D" uid="uid://b3tua3qouv2v" path="res://art/animations/weapons/dagger_sprite_sheet.png" id="2_un7sy"]
@@ -110,7 +110,6 @@ animations = [{
 [node name="Dagger" type="Node2D"]
 scale = Vector2(7, 7)
 script = ExtResource("1_5mald")
-LightSweepEnabled = false
 LightActive = 0.2
 HeavyActive = 0.5
 LightOuterRadius = 30.0
@@ -118,6 +117,7 @@ LightAngleDeg = 20.0
 HeavyInnerRadius = 15.0
 HeavyOuterRadius = 30.0
 HeavyAngleDeg = 70.0
+LightSweepEnabled = false
 
 [node name="AnimatedSprite2D" type="AnimatedSprite2D" parent="."]
 sprite_frames = SubResource("SpriteFrames_jvvga")
