+++ Game/ai_boss/scenes/new_sword.tscn
@@ -1,62 +1,93 @@
-[gd_scene load_steps=18 format=3 uid="uid://cugyiq1gej4hh"]
+[gd_scene load_steps=21 format=3 uid="uid://cqiyycxt5b10s"]
 
-[ext_resource type="Script" uid="uid://cjemx5v6p4ud" path="res://scripts/MeleeWeapon.cs" id="1_n355p"]
-[ext_resource type="Texture2D" uid="uid://diffyoy8gok5y" path="res://art/animations/weapons/sword_sprite_sheet.png" id="1_ycbs3"]
+[ext_resource type="Script" uid="uid://bv30ncfyohfjq" path="res://scripts/NewWeapon.cs" id="1_q4i4l"]
+[ext_resource type="Texture2D" uid="uid://diffyoy8gok5y" path="res://art/animations/weapons/sword_sprite_sheet.png" id="2_emisj"]
+[ext_resource type="Script" uid="uid://cpybuc7cg7rl5" path="res://scripts/attack_types/SweepingCrescentMeleeAttack.cs" id="3_x2xu1"]
+
+[sub_resource type="Resource" id="Resource_emisj"]
+script = ExtResource("3_x2xu1")
+SweepFromStartEdge = true
+SweepStepDeg = 6.0
+SweepStepDelay = 0.016
+InnerRadius = 18.0
+OuterRadius = 28.0
+AngleDeg = 70.0
+ArcCenterOffsetDeg = 0.0
+Damage = 0.0
+Cooldown = 0.15
+Windup = 0.0
+Active = 0.25
+metadata/_custom_type_script = "uid://cpybuc7cg7rl5"
+
+[sub_resource type="Resource" id="Resource_wy516"]
+script = ExtResource("3_x2xu1")
+SweepFromStartEdge = false
+SweepStepDeg = 6.0
+SweepStepDelay = 0.016
+InnerRadius = 25.0
+OuterRadius = 40.0
+AngleDeg = 120.0
+ArcCenterOffsetDeg = 0.0
+Damage = 0.0
+Cooldown = 1.0
+Windup = 0.15
+Active = 0.5
+metadata/_custom_type_script = "uid://cpybuc7cg7rl5"
 
 [sub_resource type="AtlasTexture" id="AtlasTexture_phngh"]
-atlas = ExtResource("1_ycbs3")
+atlas = ExtResource("2_emisj")
 region = Rect2(0, 48, 48, 48)
 
 [sub_resource type="AtlasTexture" id="AtlasTexture_un7sy"]
-atlas = ExtResource("1_ycbs3")
+atlas = ExtResource("2_emisj")
 region = Rect2(48, 48, 48, 48)
 
 [sub_resource type="AtlasTexture" id="AtlasTexture_mkphk"]
-atlas = ExtResource("1_ycbs3")
+atlas = ExtResource("2_emisj")
 region = Rect2(96, 48, 48, 48)
 
 [sub_resource type="AtlasTexture" id="AtlasTexture_1scc2"]
-atlas = ExtResource("1_ycbs3")
+atlas = ExtResource("2_emisj")
 region = Rect2(144, 48, 48, 48)
 
 [sub_resource type="AtlasTexture" id="AtlasTexture_mhf5r"]
-atlas = ExtResource("1_ycbs3")
+atlas = ExtResource("2_emisj")
 region = Rect2(192, 48, 48, 48)
 
 [sub_resource type="AtlasTexture" id="AtlasTexture_siec8"]
-atlas = ExtResource("1_ycbs3")
+atlas = ExtResource("2_emisj")
 region = Rect2(240, 48, 48, 48)
 
 [sub_resource type="AtlasTexture" id="AtlasTexture_rohr5"]
-atlas = ExtResource("1_ycbs3")
+atlas = ExtResource("2_emisj")
 region = Rect2(288, 48, 48, 48)
 
 [sub_resource type="AtlasTexture" id="AtlasTexture_nip8w"]
-atlas = ExtResource("1_ycbs3")
+atlas = ExtResource("2_emisj")
 region = Rect2(336, 48, 48, 48)
 
 [sub_resource type="AtlasTexture" id="AtlasTexture_06afe"]
-atlas = ExtResource("1_ycbs3")
+atlas = ExtResource("2_emisj")
 region = Rect2(0, 0, 48, 48)
 
 [sub_resource type="AtlasTexture" id="AtlasTexture_u6f1l"]
-atlas = ExtResource("1_ycbs3")
+atlas = ExtResource("2_emisj")
 region = Rect2(48, 0, 48, 48)
 
 [sub_resource type="AtlasTexture" id="AtlasTexture_xowah"]
-atlas = ExtResource("1_ycbs3")
+atlas = ExtResource("2_emisj")
 region = Rect2(96, 0, 48, 48)
 
 [sub_resource type="AtlasTexture" id="AtlasTexture_hbqb4"]
-atlas = ExtResource("1_ycbs3")
+atlas = ExtResource("2_emisj")
 region = Rect2(144, 0, 48, 48)
 
 [sub_resource type="AtlasTexture" id="AtlasTexture_veb5s"]
-atlas = ExtResource("1_ycbs3")
+atlas = ExtResource("2_emisj")
 region = Rect2(192, 0, 48, 48)
 
 [sub_resource type="AtlasTexture" id="AtlasTexture_uhpsc"]
-atlas = ExtResource("1_ycbs3")
+atlas = ExtResource("2_emisj")
 region = Rect2(240, 0, 48, 48)
 
 [sub_resource type="SpriteFrames" id="SpriteFrames_jvvga"]
@@ -116,13 +147,10 @@ animations = [{
 
 [node name="Sword" type="Node2D"]
 scale = Vector2(7, 7)
-script = ExtResource("1_n355p")
-HeavyCooldown = 1.0
-HeavyActive = 0.65
-LightOuterRadius = 30.0
-HeavyInnerRadius = 25.0
-HeavyOuterRadius = 45.0
-HeavySweepFromStartEdge = false
+script = ExtResource("1_q4i4l")
+Description = "A basic melee weapon."
+LightAttackConfig = SubResource("Resource_emisj")
+HeavyAttackConfig = SubResource("Resource_wy516")
 
 [node name="AnimatedSprite2D" type="AnimatedSprite2D" parent="."]
 sprite_frames = SubResource("SpriteFrames_jvvga")
