+++ Game/ai_boss/scripts/Main.cs
@@ -10,7 +10,7 @@ public partial class Main : Node2D
 	{
 		// Load scenes
 		beanScene = ResourceLoader.Load<PackedScene>("res://scenes/bean.tscn");
-		swordScene = ResourceLoader.Load<PackedScene>("res://scenes/sword.tscn");
+		swordScene = ResourceLoader.Load<PackedScene>("res://scenes/new_sword.tscn");
 		
 		// Instantiate Bean
 		PlayerController beanInstance = beanScene.Instantiate<PlayerController>();
