--- Game/ai_boss/scripts/MeleeWeapon.cs.uid
@@ -1,1 +1,0 @@
-uid://cjemx5v6p4ud
