--- Game/ai_boss/scripts/WeaponBase.cs.uid
@@ -1,1 +1,0 @@
-uid://cyl7i7pv7cynx
