+++ Game/ai_boss/scripts/attack_types/IAttack.cs
@@ -1,0 +1,8 @@
+using Godot;
+using System;
+
+public interface IAttack
+{
+    void Execute(Weapon weapon, Vector2 target, bool facingLeft);
+    void Interrupt(Weapon weapon);
+}
